Dev alisha sprint4

+++ app/controllers/items_controller.rb
@@ -161,25 +161,40 @@ class ItemsController < ApplicationController
 
   def destroy
     @item = Item.find(params[:id])
-    if @item.destroy
-      flash[:notice] = 'Item was successfully deleted.'
+    Rails.logger.info("Current user auth level: #{current_user.auth_level}")
+  
+    if current_user.auth_level == 2
+      if @item.destroy
+        flash[:notice] = 'Item was successfully deleted.'
+        redirect_to items_path 
+      else
+        flash[:alert] = 'Failed to delete the item.'
+        redirect_to item_path(@item) 
+      end
     else
-      flash[:alert] = 'Failed to delete the item.'
+      flash[:alert] = 'You need to be an admin to delete items.'
+      redirect_to item_path(@item) 
     end
-    redirect_to items_path
   end
 
   def set_status # rubocop:disable Metrics/AbcSize,Metrics/MethodLength
     @item = Item.find(params[:id])
     @notes = Note.where(item_id: @item.id).order('created_at DESC')
-    valid_statuses = [nil, 'Damaged', 'Lost', 'Not Available']
+  
+    if current_user.auth_level == 0
+      flash[:alert] = 'You need to be an admin or assistant to update the status of this item.'
+      redirect_to item_path(@item) and return
+    end
+  
+    valid_statuses = [nil, 'Damaged', 'Lost', 'Not Available', 'Intact']
     status = get_valid_status(item_params[:status])
+  
     if valid_statuses.include?(status) && @item.update(item_params)
       log_event(params[:id], 'status_update', "Status Updated to #{status}", session[:user_id])
       flash[:notice] = 'Item status updated successfully.'
       redirect_to @item
     else
-      flash[:notice] = 'Error updating status. Status must be nil, Damaged, Lost, or Not Available.'
+      flash[:notice] = 'Error updating status. Status must be nil, Damaged, Lost, Not Available, or Intact.'
       render :show
     end
   end
